+++ tests/migration/Administration/V6_7/Migration1757057005MailTemplateTest.php
@@ -7,8 +7,8 @@ use Shopware\Administration\Migration\V6_7\Migration1757057005MailTemplate;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelLifecycleManager;
 use Shopware\Core\Migration\Traits\MailUpdate;
-use Shopware\Tests\MailTemplateMigrationTestCase;
-use Shopware\Tests\Translations;
+use Shopware\Tests\Migration\MailTemplateMigrationTestCase;
+use Shopware\Tests\Migration\Translations;
 
 /**
  * @internal
